renamed the fileloaders package to loader, as the worldloader doesn't really load any files but just loads/initializes the world
