+++ src/main/java/com/ciryusmedia/challengenetwork/challengespluginneo/core/loader/WorldLoader.java
@@ -1,4 +1,4 @@
-package com.ciryusmedia.challengenetwork.challengespluginneo.core.fileloaders;
+package com.ciryusmedia.challengenetwork.challengespluginneo.core.loader;
 
 import com.ciryusmedia.challengenetwork.challengespluginneo.core.console.ChallengeDebugger;
 import com.ciryusmedia.challengenetwork.challengespluginneo.core.console.DebugLevel;
